feat: token submodule
