+++ src/index.ts
@@ -20,6 +20,7 @@ import {
 import {
   BalanceInterface,
   DecodedTag,
+  ExtensionOrJWK,
   OrderBookInterface,
   OrderInterface,
   PriceInterface,
@@ -33,6 +34,7 @@ import {
 } from "./faces";
 import Utils from "./utils";
 import User from "./user";
+import Token from "./token";
 
 const client = new Arweave({
   host: "arweave.net",
@@ -42,13 +44,14 @@ const client = new Arweave({
 
 export default class Verto {
   public arweave = client;
-  public wallet: "use_wallet" | JWKInterface = "use_wallet";
+  public wallet: ExtensionOrJWK = "use_wallet";
   public smartweave: SmartWeave;
   public cache: boolean;
 
   // Submodules
   private utils: Utils;
   public user: User;
+  public token: Token;
 
   /**
    *
@@ -71,177 +74,12 @@ export default class Verto {
       this.smartweave
     );
     this.user = new User(this.arweave, this.cache, this.utils);
+    this.token = new Token(this.arweave, this.wallet, this.cache, this.smartweave, this.utils);
   }
 
   // === Token Functions ===
 
-  /**
-   * Fetches the tokens listed on Verto.
-   * @returns List of token ids, names, & tickers.
-   */
-  async getTokens(type?: TokenType): Promise<TokenInterface[]> {
-    let tokens: CommunityContractToken[] = [];
-    const parsedTokens: TokenInterface[] = [];
-
-    if (this.cache) tokens = await fetchTokens(type);
-    else {
-      const contract = await this.getState(this.COMMUNITY_CONTRACT);
-
-      tokens = contract.tokens;
-    }
-
-    for (const token of tokens) {
-      if (this.cache) {
-        const data = await fetchTokenStateMetadata(token.id);
-
-        if (!data) continue;
-        parsedTokens.push({
-          id: data.id,
-          name: data.name,
-          ticker: data.ticker,
-        });
-      } else {
-        const data = await this.getState(token.id);
-
-        if (!data) continue;
-        parsedTokens.push({
-          id: token.id,
-          name: data.name,
-          ticker: data.ticker,
-        });
-      }
-    }
-
-    return parsedTokens;
-  }
-
-  // TODO
-  /**
-   * Fetches the type for a given token.
-   * @param id Token contract id.
-   * @returns The type of the token.
-   */
-  /*async getTokenType(id: string): Promise<TokenType> {
-    if (this.cache) {
-
-    } 
-  }*/
-
-  /**
-   * Fetches the latest price for a given token.
-   * @param id Token contract id.
-   * @returns The price with name & ticker, or undefined.
-   */
-  async getPrice(id: string): Promise<PriceInterface | undefined> {
-    const res = await axios.get(`${this.endpoint}/token/${id}/price`);
-    return res.data;
-  }
-
-  /**
-   * Fetches the price history for a given token.
-   * @param id Token contract id.
-   * @returns Dates mapped to prices.
-   */
-  async getPriceHistory(id: string): Promise<{ [date: string]: number }> {
-    const res = await axios.get(`${this.endpoint}/token/${id}/priceHistory`);
-    return res.data;
-  }
-
-  /**
-   * Fetches the latest volume for a given token.
-   * @param id Token contract id.
-   * @returns The volume as a number.
-   */
-  async getVolume(id: string): Promise<number> {
-    const res = await axios.get(`${this.endpoint}/token/${id}/volume`);
-    return res.data;
-  }
-
-  /**
-   * Fetches the volume history for a given token.
-   * @param id Token contract id.
-   * @returns Dates mapped to volumes.
-   */
-  async getVolumeHistory(id: string): Promise<{ [date: string]: number }> {
-    const res = await axios.get(`${this.endpoint}/token/${id}/volumeHistory`);
-    return res.data;
-  }
-
-  /**
-   * Transfer a specified amount of tokens to another wallet.
-   * @param amount The amount of tokens.
-   * @param id Token contract id.
-   * @param target The receiving address.
-   * @param tags Optional additional tags
-   * @returns The transaction id of the transfer.
-   */
-  async transfer(
-    amount: number,
-    id: string,
-    target: string,
-    tags: { name: string; value: string }[] = []
-  ) {
-    const contract = this.smartweave.contract(id).connect(this.wallet);
-    const transaction = contract.writeInteraction(
-      {
-        function: "transfer",
-        target,
-        qty: amount,
-      },
-      [
-        { name: "Exchange", value: "Verto" },
-        { name: "Action", value: "Transfer" },
-        ...tags,
-      ],
-      {
-        target,
-        winstonQty: "0",
-      }
-    );
-
-    return transaction;
-  }
-
-  /**
-   * List a new token on the exchange.
-   * @param address The ID of the token
-   * @returns InteractionID
-   */
-  async list(
-    address: string,
-    type: TokenType,
-    tags: DecodedTag[] = []
-  ): Promise<string> {
-    const contract = this.smartweave
-      .contract(this.COMMUNITY_CONTRACT)
-      .connect(this.wallet);
-
-    if (!this.validateHash(address)) throw new Error("Invalid token address.");
-
-    // TODO: do we want fees on this @t8
-    const interactionID = await contract.writeInteraction(
-      {
-        function: "list",
-        id: address,
-        type,
-      },
-      [
-        {
-          name: "Exchange",
-          value: "Verto",
-        },
-        {
-          name: "Action",
-          value: "ListToken",
-        },
-        ...tags,
-      ]
-    );
-
-    if (!interactionID) throw new Error("Could not list token.");
-
-    return interactionID;
-  }
+  
 
   /**
    * Add a new pair to the exchange.
